# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "User Book Reading Notes",
      "description": "API para registro, login e listagem de usuários",
      "version": "1.0.0"
    },
      "components":{
        "securitySchemes":{
          "BearerAuth":{
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        }      
    },
    "paths": {
        "/api/book/register": {
        "post": {
          "tags": ["Books"],
          "summary": "Registrar um novo Livro",
            "requestBody": {
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "category": {
                        "type": "string"
                      },
                      "author":{
                        "type": "string"
                      }
                    },
                    "required": ["name", "category", "author"]
                  }
                }
              }
            },
          "responses": {
            "200": {
              "description": "Livro registrado com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "category": {
                        "type": "string"
                      },
                      "author":{
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Erro ao registrar o Livro"
            }
          }
        }
      },
      "/api/book/delete/{id}": {
        "delete": {
          "tags": ["Books"],
          "summary": "Deletar livro",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Livro deletado com sucesso"
            },
            "400": {
              "description": "Erro ao deletar o livro"
            }
          }
        }
      },
      
      "/api/book/book": {
        "get": {
          "tags": ["Books"],
          "summary": "Obter a lista de Livros",          
          "responses": {
            "200": {
              "description": "Lista de livros",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string"
                        },
                        "author":{
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Erro ao obter a lista de livros"
            }
          }
        }
      } ,      
      
      "/api/user/register": {
        "post": {
          "tags": ["Users"],
          "summary": "Registrar um novo usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Usuário registrado com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "username": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Erro ao registrar o usuário"
            }
          }
        }
      },


      "/api/user/login": {
        "post": {
          "tags": ["Users"],
          "summary": "Login de usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login realizado com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Erro ao realizar login"
            }
          }
        }
      },
      

      "/api/user/user": {
        "get": {
          "tags": ["Users"],
          "summary": "Obter a lista de usuários",
          "security":[
            {
              "BearerAuth":[]
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de usuários",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "username": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Erro ao obter a lista de usuários"
            }
          }
        }
      }   
    }
  }
  